Changed stuff according to the comments. Also added module with helper functions to make other things look neater.
