+++ lib/simplifyExpression/collectAndCombineSearch/index.js
@@ -35,7 +35,7 @@ function collectAndCombineLikeTerms(node) {
   else if (node.op === '*') {
     // collect and combine involves there being coefficients pulled the front
     // e.g. 2x * x^2 * 5x => (2*5) * (x * x^2 * x) => ... => 10 x^4
-    if (checks.canMultiplyLikeTermConstantNodes.canMultiplyLikeTermConstantNodes(node)) {
+    if (checks.canMultiplyLikeTermConstantNodes(node)) {
       return multiplyLikeTerms(node, true);
     }
     const status = collectAndCombineOperation(node);
