+++ lib/simplifyExpression/collectAndCombineSearch/multiplyLikeTerms.js
@@ -1,12 +1,14 @@
 const arithmeticSearch = require('../arithmeticSearch');
 const checks = require('../../checks');
 const clone = require('../../util/clone');
+const ConstantOrPowerTerm = require('./ConstantOrPowerTerm');
 const multiplyFractionsSearch = require('../multiplyFractionsSearch');
 
 const ChangeTypes = require('../../ChangeTypes');
 const Node = require('../../node');
 
-// Multiplies a list of nodes that are polynomial or constant power like terms. Returns a node.
+// Multiplies a list of nodes that are polynomial or constant power like terms.
+// Returns a node.
 // The polynomial nodes should *not* have coefficients. (multiplying
 // coefficients is handled in collecting like terms for multiplication)
 function multiplyLikeTerms(node, polynomialOnly=false) {
@@ -15,7 +17,7 @@ function multiplyLikeTerms(node, polynomialOnly=false) {
   }
   let status;
 
-  if (!polynomialOnly && !checks.canMultiplyLikeTermConstantNodes.canMultiplyLikeTermConstantNodes(node)) {
+  if (!polynomialOnly && !checks.canMultiplyLikeTermConstantNodes(node)) {
     status = arithmeticSearch(node);
     if (status.hasChanged()) {
       status.changeType = ChangeTypes.MULTIPLY_COEFFICIENTS;
@@ -39,7 +41,7 @@ function multiplyLikeTerms(node, polynomialOnly=false) {
 }
 
 function multiplyPolynomialTerms(node) {
-  if (!checks.canMultiplyLikeTermPolynomialNodes(node) && !checks.canMultiplyLikeTermConstantNodes.canMultiplyLikeTermConstantNodes(node)) {
+  if (!checks.canMultiplyLikeTermPolynomialNodes(node) && !checks.canMultiplyLikeTermConstantNodes(node)) {
     return Node.Status.noChange(node);
   }
   const substeps = [];
@@ -58,7 +60,7 @@ function multiplyPolynomialTerms(node) {
   // STEP 2: collect exponents to a single exponent sum
   // e.g. x^1 * x^3 -> x^(1+3)
   // e.g. 10^2 * 10^3 -> 10^(2+3)
-  if (checks.canMultiplyLikeTermConstantNodes.canMultiplyLikeTermConstantNodes(node)) {
+  if (checks.canMultiplyLikeTermConstantNodes(node)) {
     status = collectConstantExponents(newNode);
   }
   else {
@@ -101,11 +103,11 @@ function addOneExponent(node) {
   let change = false;
 
   let changeGroup = 1;
-  if (checks.canMultiplyLikeTermConstantNodes.canMultiplyLikeTermConstantNodes(node)) {
+  if (checks.canMultiplyLikeTermConstantNodes(node)) {
     newNode.args.forEach((child, i) => {
-      if (!checks.canMultiplyLikeTermConstantNodes.getExponentNode(child)) {
+      if (!ConstantOrPowerTerm.getExponentNode(child)) {
         newNode.args[i] = Node.Creator.powerTerm(
-          checks.canMultiplyLikeTermConstantNodes.getBaseNode(child),
+          ConstantOrPowerTerm.getBaseNode(child),
           Node.Creator.constant(1));
 
         newNode.args[i].changeGroup = changeGroup;
@@ -148,10 +150,10 @@ function addOneExponent(node) {
 function collectConstantExponents(node) {
   // If we're multiplying constant nodes together, they all share the same
   // base. Get that from the first node.
-  const baseNode = checks.canMultiplyLikeTermConstantNodes.getBaseNode(node.args[0]);
+  const baseNode = ConstantOrPowerTerm.getBaseNode(node.args[0]);
   // The new exponent will be a sum of exponents (an operation, wrapped in
   // parens) e.g. 10^(3+4+5)
-  const exponentNodeList = node.args.map(p => checks.canMultiplyLikeTermConstantNodes.getExponentNode(p));
+  const exponentNodeList = node.args.map(p => ConstantOrPowerTerm.getExponentNode(p));
   const newExponent = Node.Creator.parenthesis(
     Node.Creator.operator('+', exponentNodeList));
   const newNode = Node.Creator.powerTerm(baseNode, newExponent, null);
